client/requirements.txt
click>=8.0.0
requests>=2.28.0
pyyaml>=6.0
python-dateutil>=2.8.0
